+++ ChangeLog.md
@@ -15,6 +15,9 @@ v5.4.1p1
   - #602 Sisimai partially supports the media types described in RFC6533 such as `message/global`.
   - #606 #607 Fix the index out of range bug in multipart blocks.
   - #605 #608 Implement `Sisimai::Lhost::Mimecast`.
+  - #566 #548 Remove the following accessors and JSON keys of `Sisimai::Fact` deprecated at v5.2.0.
+    - `smtpagent` (use `decodedby` instead)
+    - `smtpcommand` (use `command` instead)
 
 v5.4.1
 ---------------------------------------------------------------------------------------------------